requirements.txt
celery==5.2.7
certifi==2023.7.22
coverage==7.2.6
iniconfig==2.0.0
opentelemetry-api==1.19.0
opentelemetry-sdk==1.19.0
opentelemetry-exporter-jaeger==1.19.0
opentelemetry-exporter-otlp-proto-grpc==1.19.0
opentelemetry-semantic-conventions==0.40b0
packaging==23.1
pluggy==1.0.0
project-paths==1.1.1
pymongo==3.12.1
pytest==7.3.1
pytest-mock==3.10.0
python-dotenv==1.0.0
requests==2.31.0
requests-mock==1.10.0
toml==0.10.2
tomli==2.0.1
urllib3==2.0.7
